(PUP-4567) Update acceptance test
Prior to this change, only a single test could be run, because `puppet`
was only installed at the start of the test suite. Change to a pattern
where every test sets up `puppet`, then cleans up after the test.

Update the existing acceptance test to expect the `puppet-agent`
package with the `puppet` and `mcollective` services running. Also
don't expect idempotency to work yet, because we switch from `puppet`
to `puppet-agent` mid-test.

Add new acceptance tests for upgrading without starting services, and
upgrading when mcollective is configured.