(PUP-4567) Add master/agent upgrade test
Prior to this, no acceptance test verified that the upgrade module
worked when run in a master/agent configuration, only masterless.

Add a test verifying master/agent behavior. This includes lots of hacks
due to BKR-273, BKR-254, waiting on Puppet Server compatibility with
Puppet 3 agents in SERVER-595, and a release of Puppet 3.8.1 packages.