(maint) Only match hostname against mco ping output
Previously, the tests failed when run against the vmpooler, because the
vmpooler returns a dynamically generated fqdn as the `default` node's
hostname, and that's what `Beaker::Host#to_s` returns. However, `mco ping`
only outputs the hostname:

    izem5qqz7rbrm0u.delivery.puppetlabs.net (agent) 23:08:26$ mco ping
    izem5qqz7rbrm0u                          time=173.65 ms

This commit calls the `Host#hostname` method, extracts the hostname
component of the fqdn, and matches against the hostname and optional
domain component.