(PE-10132) Use source parameter if specified
Previous to this commit, the only platform to take advantage of the
source parameter on the `puppet_agent` class was windows, leaving Debian
and Redhat with a hardcoded repository source.
This commit adds support for both debian and redhat platforms to use the
source parameter. If the parameter is not specified, it will fall back
to using their respective puppetlabs.com repositories.