(PE-10132) Add support for pe_repo as a source
This commit adds pe_repo support for users in a Puppet Enterprise
environment. In a typical PE install, the PE master will be assigned the
pe_repo class and will host the packages, giving us a safe default. In
the event that they have changed either the server or repo_dir
parameters, then they must pass in the source parameter.