(PE-10132) Change way is_pe is determined
Previous to this commit, determinig if the module was running in a PE
environment was to use the fact provided by stdlib `$::is_pe`. However
with the AIO Puppet agent shipping with the next versio of PE, the fact
will no longer work (in it's current form).
This commit refactors usage of it to params.pp and calculates it using
both the `$::is_pe` fact and stdlibs `is_function_available` and
checking for a function provided by the Puppet Enterprise module.