(PUP-4849) Make puppet_agent ensure platform type
In the future we may want to add acceptance tests that verify upgrading
a puppet-agent package to a newer version. For that case, using type=aio
makes sense as part of config.

The call to `install_puppet_on` should set the host to the appropriate
type, but the helper `remove_defaults_on` is broken (see BKR-395) when
a type is set in the config file; replace with `configure_defaults_on`.

[skip ci]