(PE-11167) Reset puppet.conf classfile setting
This commit adds the puppet.conf classfile setting to the list of
settings to be reset during agent ugprade. This is needed primarily for
PE users as the path to that file was set to a different directory at
install time. With the new AIO agent, that file location has changed. To
bring things closer to parity and to continue configuring PE users
mcollective with mco inventory support, `classfile` is no longer being
managed at install time in PE. This commit brings it full circle by
essentially undoing PE's previous one-time configuration of the
`classfile` setting.