(PE-11168) Use specific agent version in PE
Previous to this commit, the windows agent upgrade code would always use
the latest agent version available. In PE, the agent version should be
tied to the master's. However Windows does not have a curl like function
in powershell v2 that supports self signed certificates. This commit
overrides the source url if not specified and the user is in a PE
environment to use the PE download url, specifing the agent version,
sourced from the compiling master.