(PE-10915) Base master_agent_version on pe_compiling_server_aio_build()
We were determining the master's agent_version based on
file('/opt/puppetlabs/puppet/VERSION'). Since we need to refer to this
in more than one class now, and since it is only used in PE, it makes
sense to move it to the params class and to base it off of the
pe_compiling_server_aio_build() function which inspects the same file.
This way we're not duplicating the function across modules, and we're
not mocking the builtin file() function specifically in our specs.