(PE-12001) Ensure pe-puppet/mcollective services stopped
Removing the PUPpuppet and PUPmcollective packages does not stop the
running services.  I've added service resources for both of the old PE
services to ensure they are stopped.  However, it is also critical that
they are stopped as the very last act of the catalog, because stopping
the pe-puppet service, at least on Solaris, terminates the run in mid
progress.  Originally I had the services stopped in the prepare stage,
but had to reposition them.