(PE-13608) Do not convert windows file resource to RAL catalog
Before this commit when rspec-puppet would attempt to use the
'that_comes_before' matcher on a POSIX system with the file resource
when the file resource contained a windows path the test would fail.
This is because when puppet converts the catalog to the RAL catalog type
it pokes around in the actual types themselves. Since this is done on a
POSIX based system puppet attempts to validate the absolute path in the
file resource as a POSIX file path and fails since it is not a valid
absolute path.

This commit replaces the call to create the relationship graph to
determine where the windows file resource is at with a more basic test
to assert that the 'before' parameter is filled out on the resource
instead.