(PE-12299) Start services after puppet run
Before this commit when upgrading solaris 11 agents the services would
be in a maintenance state where puppet was unable to start them. This
commit adds in a script which waits for the currently running agent run
to stop and then calls puppet to ensure that puppet and mcollective are
running so that users do not have to login to the nodes to perform these
steps by hand.

This behavior is currently what windows is doing to start puppet and
mcollective agents when upgrading with the puppet_agent module.