(PE-11531) Change `version` to `package_version`
Prior to this commit we added the parameter `version` to specify
the `puppet-agent` version that you wanted to install. That could
be confusing, as some users might think of just the puppet version,
and not the puppet-agent package version.
This commit updates the parameter `version` to be `package_version`,
which, hopefully, will be more clear to users.

Additionally, this updates the version check regex to match the
actual version numbers formats that we can put on puppet-agent packages.