(PE-10956) Do not manage service on Windows
Prior to this commit we were attempting to manage the
puppet/mcollective services on Windows systems, which
resulted in errors in the puppet run due to the MSI being
in the middle of configuring those services.
This commit removes management of the puppet/mco services
on windows systems, because the MSI is handling that already
and attempting to manage it in puppet just creates a race
condition.