Fix SLES11 GPG key import issue
Previously, the exec to import the Puppetlabs GPG key made a call out to the
gpg binary to gather the hex value of the Puppetlabs GPG key. This call works
from the console with an ENV, but it doesn't work when automated because it
cannot find the relative homedir (~/.gnupg). To work around this, explicitly
set the value for root's .gnupg homedir in the unless command and expose it as
a variable should we need to deal with it later.