(PE-10956) Add start /wait to windows upgrade
Prior to this commit our call to msiexec would not block until
the msi finished. This resulted in any commands dependent on
puppet that were run after a puppet_agent upgrade puppet run would
fail due to puppet still being in the middle of upgrading.
This commit adds a `start /wait` to the msiexec so that, hopefully,
we will block until the MSI finishes running and puppet is actually
upgraded.