(PE-10956) Pass in configured server to windows MSI
Prior to this commit when re-installing the MSI on windows systems,
it would automatically revert the Puppet server setting to whatever
was defined when puppet was initially installed. If that value was
configured to be 'puppet' because no argument was given (as is the
case in beaker) then it would default back to 'puppet', even though
we had set the server to the appropriate master with beaker post-MSI
install.
This commit gets around that surprising behavior by always providing
the currently configured Puppet server value to the MSI.