(PE-10956) Create puppet_master_server fact
Prior to this commit we were setting the agent's `server` setting
to whatever `settings::server` was set to, which in a compile master
scenario, would be the MoM. This certainly isn't what a user would
want to have happen in production.
This commit creates a `puppet_master_server` fact populated with the
value of `Puppet.settings['server']` on the agent and uses that as the
argument to the windows MSI upgrade so we don't override the existing
config value with the default of `puppet` or whatever was written to
the registry on install.