(maint) Re-enable open-source upgrades from 3.x
The prior commit to enable puppet-agent to puppet-agent upgrades set an
expectation that a target version would always be required, either from
the master or explicitly specified as a parameter. That broke the
previous open-source upgrade behavior when upgrading from Puppet 3.x.

Restore that behavior by allowing the target version to be unspecified
if not on Puppet Enterprise and upgrading from Puppet 3.x. Also add the
new dependency puppetlabs-transition for acceptance testing.