(PE-11531) Don't remove packages if package_version undef
Prior to this commit we were removing the puppet-agent package
if the `package_version` happened to be unset, due to the
behavior of `versioncmp`. This is obviously not what we want to
have happen.
This commit updates all of the calls that compare `package_version`
with `aio_agent_version` with `versioncmp` to be gated on
`package_version` not being undef, thus preventing the packages
from being removed on every run.