(PE-11531) Centralize check for aio upgrade
Prior to this commit we had the same AIO upgrade expression scattered
in 3 different places with slightly different syntax. Also, we did not
account for the fact that `aio_agent_version` is a PE only fact.
This commit creates a `$aio_upgrade_required` var in the main class
that will 1) always be true if not PE and `package_version` is set 2) be
true if `aio_agent_version` and `package_version` are set and
the `aio_agent_version` is less than (-1) the `package_version`.