(PE-11531) Fix upgrade issue with dev builds
Prior to this commit the module would always upgrade with dev
builds because they have a git SHA appended to the end, which
`versioncmp` considers to be greater than the version reported
by `aio_agent_version`, which lacks the extra git sha characters.
This commit updates the update check to strip the git sha when
comparing the `aio_agent_version` to the requested `packace_version`.