(MODULES-3304/PE-15256) Fix Windows 2008 upgrades
Prior to this commit the path being used to locate the local
prepared copy of the windows MSI had a bug where one area of
the module would use a forward slash, and the other would use a
backslash. Additionally, the install class did not depend on the
prepare step's file resource that is created and instead used the
`defined` function to check for the file.
This commit updates the slashes to be consistent and adds an
explicit require on the File resource that prepares the local
copy of the MSI.