126 - Explicitly setting environment to nodes current environment
We have to shell out to puppet config print to preserve environment
config. This gives us how the agent is configured, not how the enc or --environment
override is set, this 100% ensures that the msi installer when it writes
the environment configuration value is how the agent was configure
pre-upgrade.