(PE-16317) Disable client SSL verification on Xenial
This commit disables client SSL verification on Xenial. We had to
do this in order to work around an issue where client ssl seems to
not be compatible with the SSL certs that PE uses. The root cause of
this change, and why it only occurs in Xenial, is unknown at this time.