(MODULES-3657) Fix waiting on Windows during upgrade
Previously the script was using the timeout.exe application however this
attempts to redirect STDIN and during a headless operation, for example during
an actual puppet run, this throws an error and does not wait the proper time.
This commit changes the wait method to use ping.  Ping waits 1 second between
attempts therefore a simple `ping 127.0.0.1 -n 6` will pause for 5 seconds.

This commit also changes the location of the PID file slightly.  Previously it
used `%TEMP%` however this variable can be removed or be in an unexpected place.
Instead the PID file is always placed in the same directory as the
script (`%~dp0`).

To assist in debugging upgrade failures the install batch file also outputs
all environment variables prior to running.