(PE-17157) Do not attempt to unzip tarball when upgrade is unneeded
Before this commit if you were on solaris with puppet aio and ran puppet
agent -t after classifying the node with the puppet agent module the
manifest would fail to apply because there was no logic which gated us
from unzipping the agent tarball if an upgrade was not needed.

This commit verifies that the version of puppet agent on the master is
the same on the agent before attempting an upgrade.