(PE-17663) update logic for solaris 11 package name
This commit updates the logic for cleaning a vanagon package version
name to be compatible with the IPS requirements for solaris ll p5p
package naming. It adds the same logic found in vanagon itself for
removing any leading zeros from the dotted segments of the version
string.

Prior to this commit, leading zeros in a version segment were left
in the package name which violates p5p naming rules. This caused
vanagon builds containing such a version to fail the install
process.

Ref:
[vanagon ips version code](https://github.com/puppetlabs/vanagon/blob/master/lib/vanagon/platform/solaris_11.rb#L94-L107)
[Solaris 11 package versioning](http://www.oracle.com/technetwork/articles/servers-storage-admin/ips-package-versioning-2232906.html)