(maint) add task for generating nodesets for testing
This commit adds a rake task (copied from puppetlabs-apt) that uses
beaker-hostgenerator to generate a nodeset compatible with our vmpooler. This
enables local testing with the vmpooler, in addition to the default testing
using vagrant. Note that we still need to tell beaker to look for our nodeset
using BEAKER_setfile=<nodeset>. Credit to @DavidS for the original contribution
to puppetlabs-apt.

Signed-off-by: Moses Mendoza <moses@puppet.com>