(MODULES-3872) add acceptance test for manage_repo param
This commit adds tests to the acceptance test suite in `class_spec.rb` for
confirming that the manage_repo parameter does, in fact, do what it says.
Specifically, we test that puppet created the repo by applying the
'puppet_agent' class with no params (implying default value of true for
`manage_repo`) and then specifically ensuring the repo present with puppet and
catching (failing) on any resultant changes. Then we remove the repo
configuration and apply the puppet agent class again but with `manage_repo =>
false`, and ensure that the repo configs are not re-created (again by catching
changes).