(maint) Add MAINTAINERS file
This commit adds a MAINTAINERS file to the puppetlabs-puppet_agent repo in the
format described by github.com/puppetlabs/maintainers. It is created initially
with the two individuals currently listed in the repository README. This change
is to bring this repo inline with others at Puppet following the maintainer
paradigm.

Signed-off-by: Moses Mendoza <moses@puppet.com>