(maint) convert install_puppet.bat to CRLF
Use CRLF instead of just LF for the template. It is not
known if using LF will cause issues with Batch files in
any versions of Windows. It is thought that running a
batch with LF should work, but it is possible that there
may be some issues somewhere. It's typically better to be
on the safe side and ensure the file is CRLF. Newer versions
of Puppet should leave the EOL format the same.