(MODULES-4030) Always prepare package
Prior to this commit we attempted to "save" the master from having to
perform checksum logic on the puppet-agent tarball by only attempting
to checksum if the master_agent_version != aio_agent version. Which
would be true if
A) The agent was pre-aio (< 3.8)
B) The agent needed to be upgraded
but not true if the agent was already up to date previous to the
puppet_agent module being installed. If it was already up to date,
this prepare step would not occur, and the module would later attempt
to install using a tarball that had not been copied over.

This commit removes the master_agent_version != aio_agent version
check and replaces it with just checking for `is_pe`, since we
depend on the `pe_build_version` function. It also specifies
the `sha256lite` checksum method should be used. This seems like a
decent compromise to both save the puppetserver work as well as
make reasoning about this module easier (if the prepare class is
included, then the agent will be prepared on disk.)