(MODULES-4092) Install solaris 10 package per-zone
Solaris 10 packaging, when combined with zones, is super weird. The
Puppet Enterprise frictionless install has historically made sure that
the agent is installed per-zone. However, the `puppet_agent` module
did not. This leads to conflicts when upgrading agents - once the
Global zone is updated, trying to upgrade the other zones will fail,
since the `puppet-agent` package is "already installed".

This changes solaris 10 packaging to be per-zone, which matches how we
have historically done that deployment.