(maint) Fix config for Jenkins
Module config got out-of-sync with expectations in our Jenkins setup.
Fix it to use PUPPET_GEM_VERSION to select Puppet version for testing,
and change the default to a more recent version of Puppet.

Also fix an unquoted variable that can be undef failing Puppet's type
checking for `versioncmp`, and adapt to a change in puppet-rspec.