(MODULES-4214) Add additional installation parameters during upgrade
Previously it was not possible to pass additional installation parameters during
an upgrade process.  In the case of MSI files this could be sensitive passwords
or other public MSI properties.  In the case of package resources, additional
options could not be added to the package resource.

This commit adds a parameter called `install_options` which mirrors the
`install_options` parameter from the package resource and then passes this down
so that it is invoked on installation (either by a package resource or in the
batch file to install the MSI)