(maint) Add sync.yml file for Module Sync
Previously this module was not being managed by module sync, the preferred tool
by the Modules Team at Puppet.  This commit adds a sync.yml file which will
allow this module to be managed by module sync.  The sync.yml file emulates the
existing module configuration as much as possible.