(maint) Speed up bundle install on Travis
Previously, Travis CI runs were throwing warnings and failing to bundle in a
timely fashion because it was using the ruby gems API and was requesting too
much information.  This commit instructions bundler will download the full index
and compute the dependencies locally.  This commit also adds a `before_script`
step to output the Gemfile.lock file for debug purposes.