(maint) Remove require to vendored gem
PUP-7114 moved the semantic vendored library which broke this require.
We shouldn't be referencing the vendored libraries at all since their
API's can change without warning (as has happened in the 4.9.0 version
of puppet).

Since nothing important uses this library in the Rakefile it is safe to
remove this line.