(maint) Add pid file to Solaris upgrade
This makes it easier to test when the upgrade is still in-progress, and
mirrors how the Windows upgrade is handled. It's either this or we grep
ps for running upgrades. Currently used in upgrade acceptance tests.

This pattern - laying down a file in /tmp/puppet_agent_upgrade.pid -
should be followed for any scripts used to perform async upgrades to
maintain testability with our current test suite.