(MODULES-6686) Fix puppet_agent_spec tests for Windows
Previously many tests in classes/puppet_agent_spec.rb were not running on
Windows due to a bug in rspec-puppet-facts.  However in v1.9.0 of the gem it
now adds better windows support, but the puppet_agent_spec.rb began running
tests for Windows which then failed.  This commit modifies the test fixtures and
logic to only test things which make sense on Windows.

Also https://github.com/mcanevet/rspec-puppet-facts/issues/68 meant that the
order of the testing had to be determinstic, ensuring that Windows platforms
were tested last.