(PA-1887) Fix osfamily::darwin OSX version regex
The previous regex used a character class instead of a union to
match the major Mac OSX product version, which would match values
such as "10.14", "10.15". This meant that the puppet_agent module
would think these are supported platforms, and proceed further with
the installation (failing at a different point, which would produce
a harder-to-read error message). This commit fixes things so that
non-supported Mac OSX platforms correctly fail in the osfamily::darwin
class with a readable error message.