(MODULES-5230) Do not manage PA version on PE infra nodes
Prior to this commit, we would manage the puppet-agent version
on any node that was classified with the puppet_agent class.

After this commit, we now check to see if the node is a PE infra
node and if so print a message saying we're not going to do anything
because PE infra nodes have a specific version of puppet-agent
on them for a reason.

You would not, for instance, want your compile master to get its
agent downgraded right after you completed an upgrade of your PE
installation.