(MODULES-5230) Use legacy fact variables
Previously in commit a6587665484 the puppet_agent module was changed to not do
anything if ran on a Puppet Master.  However the fact used uses the "newer"
$facts variable which does not exist in Puppet 3.8, which the 1.x series
supports.  This commit changes the fact to the legacy variable name.  Note that
this commit should NOT be merged into the 2.x series as this no longer supports
3.x and therefore is not an issue.