(maint) Pin Beaker on old ruby versions
Previously travis was throwing Gem dependency errors on older ruby versions due
to beaker, and it's associated gems.  This commit sets the BEAKER_VERSION
environment variable which pins the Beaker gem to version 3.0 so that bundle
install complete.  Note this does not affect the actual testing as beaker is
not used, only the bundling was failing.