(MODULES-5230) Conditionally test on Puppet 3.x
Previously we would test the fact pe_server_version however in Puppet 3.x the
rspec-puppet helper is not working.  This commit modifies the test to only run
on Puppet 4.x tests and that test coverage is enough to cover Puppet 3.x.  This
commit should not be merged up to the 2.x series.