(MODULES-7580) Fix PowerShell task for puppet version on 32bit
Previously the powershell based version task was failing on 32bit windows
operating systems due to passing in a null path to Test-Path.  This commit
modifies the detection to be far more tolerant of errors and instead detects if
the registry entries are null instead of using Test-Path.

This commit also cleans up the code as per PS Script Analyzer rules;
* Remove use of aliases
* Prefer `$null -eq ...` instead of `... -eq $null`