(maint) Disable ssh host-key-check in config
This branch is pointed to by the experimental pipeline for running puppet_agent task acceptance tests. The necessary change in bolt is pointed to in the Gemfile. Specifically removing host key checks as the Jenkins controller has many entries in ~/.ssh/known_hosts that will cause sporadic failures for test targets.