(maint) Convert solaris_install.sh.epp to an ERB template
We do this to prepare for the work in MODULES-7698. Specifically,
MODULES-7698 requires us to upgrade our OSX agents in the same way we
upgrade our Solaris 10 agents, which is by running the upgrade in the
background after the Puppet run is finished. The OSX upgrade script will
have a similar structure to the Solaris 10 upgrade script, with the only
difference being how the agent is installed. Thus, we would like to
generalize some of this structure in a separate commit. Our generalized
script would have a spot for us to insert our platform-specific agent
installation code. "Inserting" our platform-specific agent installation
code really means rendering the corresponding script template inside our
generic installation script. There's not an easy (or clean) way to
support rendering another EPP script inside an EPP script. However, it
is easy to do this with ERB templates. Thus, this commit converts our
solaris_install.sh.epp script to an ERB script so that we are able to
generalize the common structure between our OSX and Solaris upgrade
scripts in future commits.