(maint) Add do_install.sh.erb script
This script is a generic agent installation script used as a template
for upgrading the agent on non-upgradeable platforms like Solaris 10 and OSX.
We refactor our Solaris 10 agent upgrades using this template here. We
will refactor our OSX agent upgrades using this template in a later
commit as part of MODULES-7698.