(maint) Explicitly upgrade when version is not specified for yum
The tasks acceptance tests will install puppet-agent 5.5.3 from the `puppet5` collection by setting both the `collection` and `version` task parameters. Then the task is run again with only the `collection` parameter set to `puppet5` with the expectation that the latest version from the collection will be installed. For fedora 26 when `yum install puppet-agent` was run the upgrade happened as expected. For fedora 27 yum must explicitly upgrade puppet agent in order to have the desired result.