(BOLT-915) Use facts module to query for target platform
Much of the puppet_agent task is logic to determine the operating systems and version of the target. This is duplicated by the shell implementation of the facts task. The addition of multi-file task support allows using the facts module bash.sh implementation code. The bash.sh implementation has been modified to return platform and version when invoked with the positional parameter `platform` or `release'. This commit utilizes the code in the facts module to determine platform and version.