(PA-2282) Add Docker workflow for iterative development
This commit decreases the feedback cycle time to enable iterative
development. The gist is that a developer can use Docker to test
out upgrades on a machine.

The `docker/bin/upgrade.sh [<PLATFORM>] [<BEFORE>] [<AFTER>]`
command builds a container on the `<PLATFORM>` platform that has
the `<BEFORE>` version of puppet on it and adds the `puppet_agent`
module. It then runs `puppet apply` in the container such that the
`<AFTER>` version of puppet exists afterward. This script ends with
a shell inside the container.

Available platforms are Ubuntu Xenial (`ubuntu`) and CentOS 7
(`centos`).

Also available is `docker/bin/versions.sh [<PLATFORMS>]`, which
will output which versions of puppet are available on given
comma-separated `<PLATFORMS>`.

For full details, see docker/ubuntu/Dockerfile or
docker/centos/Dockerfile.