(MODULES-8445) Fix collection determination
By default, this module installs the most recent version of puppet-agent
that's compatible with the master's version, using puppet collections to
do so.

Previously, the name of the default puppet collection was determined
based on the installed version of PE, which was technically correct for
PE installations, but did not help outside of PE. For open source users,
it incorrectly selected 'PC1' as the collection name regardless of which
version of the agent was installed on the master.

This derives the collection name from the aio_agent_version fact on the
master instead, as follows:

- Agents in the 1.y.z series (containing puppet 4) still get `PC1`
- Agents in the 5.y.z series (containing puppet 5) now get `puppet5`
- Agents in the 6.y.z series (containing puppet 6) now get `puppet6`
- Later agents will get 'puppet', which always defaults to the latest
  release (this would be puppet6 right now)
- The default, absent any aio_agent_version information, is still to
  install from `PC1`