(maint) Update task_acceptance tests to use updated dependencies
Previously bolt was pinned back to 1.1.x due to an issue in puppetlabs/bolt#824 which was fixed in 1.9.0. There has also been a new release in the facts module we would like to be testing with. Finally the task acceptance test invocation has been modified to be more succinct and to match other modules that use conditionally require bolt to test module content.