(MODULES-4730) Do not pass environment on upgrade
Modern agent installs should not rely on environment being specified in
puppet.conf, preferring instead that environment assignment is defined
centrally. In the event puppet.conf does not already have an
environment= line, it is important NOT to add one. Passing
PUPPET_AGENT_ENVIRONMENT to the MSI will have the effect of adding this
line to puppet.conf. Therefore, only do it if an environment line
already exists.