(MODULES-8319) Update service management to ignore mco after puppet6
Packages from the puppet6 collections (and moving forward from that) we no
longer carry the mcollective service. So for upgrades to collections other
than puppet5 and PC1 we should ignore the mcollective service.

This commit updates the service logic to use a new $_service_names parameter
that is updated to exclude mcollective with anything that is not collections
puppet5 or PC1