(MODULES-4730) Do not pass the agent environment during MSI installs
Since the MSI installer already preserves the environment of an existing
agent install, there's no need to specify it. Additionally, when there is
no 'environment' key present in puppet.conf, this would set the environment
key to the default value 'production', which can cause nodes to be
incorrectly classified after upgrades.