(maint) Refactor structure of package sources
This commit performs a major refactor of the structure of identifying
and using sources for the packages to upgrade. Before this commit,
sources were identified in a myriad of different ways with no
consistency.

All source identification now happens as the first action in the
osfamily manifests. Sources are identified in these manifests and
then used to pull packages sources (if necessary) or identify where
the osfamily manifest downloads repos.

Parts of the init manifest that were previously calculating things
like package name have been moved to the osfamily manifests as well.