Fix unnecessary changes in loadbalanced envs.
In environments with multiple compile masters behind a load balancer,
using `$::servername` could cause Puppet to constantly update the repo
configuration based on the compile master that happens to be compiling
the catalog at that time.

Changing the value to `$::puppet_master_server` uses the server value in
puppet.conf, which in most cases should be the load balanced point.

Update `puppet_master_server` fact to support environments using
`server_list` in their puppet.conf file instead of `server`. If `server`
is not defined, the first master defined in `server_list` will be used.