(MODULES-8821) Update win install to use production environment
The windows installer uses the puppet command from the CLI to find the
statedir to put install state files in. When the default environment for
puppet is set to something other than production the command fails on older
agents (See: https://tickets.puppetlabs.com/browse/PUP-6739).

This commit simply updates the invocation of the puppet command to force the
production environment when finding the statedir.