(MODULES-9497) install_puppet.ps1 stale .pid file
When a agent upgrade fails for some reason, the
puppet_agent_upgrade.pid file is not deleted, and the next time
the install_puppet.ps1 runs, it fails again and again as it cannot start
the installation because of the stale .pid file.

Now, the install_puppet.ps1 checks if the PID inside the .pid file is
still running, and if not, it will replace the PID inside the file and
continue the installation.